requirements.txt
pyzmq>=22.3.0
ipython>=8.0
pytest>=7.0
